Problem 2 in python
Interesting to see that the castings are functions, argv[0] is the
script name, and the syntaxis for generators.

I would like something in between python and ruby. 'Magically'
converting any function with a yield to a generator, sounds weird.

I'm starting to like the indentation. Also the automatic whitespace in
the print with multiple arguments. In ruby the FS is the new line.