Finishing problem #4
I'm keeping the 3 approaches for history. Turns out that generators
(option 2) are faster than the alternative.

I thought it would be faster by a long shot but it's only by a bit.

I also thought option 3 could be faster (though it would require more
memory) but seems to be as fast if not worse than option 2.